+++ src/point.hpp
@@ -4,10 +4,10 @@
 
 class Point {
     public:
-        float x = NAN, y = NAN, z = NAN;
+        double x = NAN, y = NAN, z = NAN;
         
         Point() {};
-        Point(float coord_x, float coord_y, float coord_z) {
+        Point(double coord_x, double coord_y, double coord_z) {
             x = coord_x;
             y = coord_y;
             z = coord_z;
