+++ src/segment.hpp
@@ -25,6 +25,6 @@ class Segment {
         bool is_nan() const;
         void print() const;
         bool is_on_same_line(const Segment& segment) const;
-        float distance_between_lines(const Segment& segment) const;
+        double distance_between_lines(const Segment& segment) const;
         Point segments_intersection(const Segment& segment) const;
 };
